Create a base class for saving/loading json settings
